bbsolr-4.8.1-news: multivalued field support for PerSegmentFaceting (part 1/2)
Summary:
This revision adds support for multivalued fields to `PerSegmentSingleValuedFaceting`.
In the next part we move this to a new class - `PerSegmentFaceting` and revert `PerSegmentSingleValuedFaceting` to its original code.

I've tried to keep refactoring of existing code out of this revision in general, in order to keep the diff simple.

Test Plan: -

Reviewers: dceccarelli4, raiyenga

Reviewed By: dceccarelli4

Differential Revision: https://all.phab.dev.bloomberg.com/D871018